+++ java/com/google/cm/mrp/WorkerPullWorkService.java
@@ -36,6 +36,9 @@ import java.security.Security;
 import java.time.Duration;
 import java.util.Optional;
 import javax.inject.Inject;
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.core.config.Configurator;
 import org.conscrypt.Conscrypt;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -67,6 +70,20 @@ public class WorkerPullWorkService extends AbstractExecutionThreadService {
   @Override
   protected void run() {
     logger.info("Worker starting run.");
+
+    try {
+      startupConfigProvider
+          .getStartupConfig()
+          .loggingLevel()
+          .ifPresent(
+              logLevel -> {
+                logger.info("Setting MRP log level to: {}", logLevel);
+                Configurator.setAllLevels(
+                    LogManager.getRootLogger().getName(), Level.getLevel(logLevel));
+              });
+    } catch (Exception e) {
+      logger.warn("Could not set logging level", e);
+    }
 
     if (startupConfigProvider.getStartupConfig().conscryptEnabled()) {
       // Initialize conscrypt