+++ RELEASES.md
@@ -1,5 +1,15 @@
 # Confidential Match Source Code Releases
 
+## February 13, 2025
+
+* Version: v1.69.0
+* Lookup Service
+  * Image Digest: sha256:317b4f2034668087a30300cdc095e166af1ce6a36091a4cdffd64736778aa493
+  * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/lookup_server_gcp_signature@sha256:30ea94d8b2461256024acf61c7e966438d4b8037e274b062c5e29fe60e9e71fb)
+* Match Request Processor
+  * Image Digest: sha256:cbbd7a03a80ea9a68d790094c4775d640732976e710111eb50d4ba4878572e06
+  * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/mrp_app_gcp_signature@sha256:8096fd28e26d85e5b1b20d561704ba54fcddf6f294888faf1266508e0bc14c63)
+
 ## February 5, 2025
 
 * Version: v1.61.0